requirements.txt
flask
requests
flask-limiter
cryptography
openai
python-dotenv
